+++ docs/source/conf.py
@@ -5,8 +5,8 @@
 
 # -- Path setup --------------------------------------------------------------
 
-from importlib.metadata import version
 import configparser
+import json
 import pathlib
 import sys
 
@@ -30,8 +30,24 @@ rst_prolog = f".. |author| replace:: {author}"
 # -- Special handling for version numbers ------------------------------------
 # https://github.com/pypa/setuptools_scm#usage-from-sphinx
 
-release = version(project)
+gh_org = "BCDA-APS"
+project = project
+release = apstools.__version__
 version = ".".join(release.split(".")[:2])
+
+# fmt: off
+switcher_file = "_static/switcher.json"
+switcher_json_url = (
+    "https://raw.githubusercontent.com/"
+    f"{gh_org}/{project}/"
+    "main/docs/source"
+    f"/{switcher_file}"
+)
+switcher_version_list = [
+    v["version"]  # to match with ``release`` (above)
+    for v in json.load(open(switcher_file))
+]
+# fmt: on
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -57,10 +73,20 @@ exclude_patterns = ["**.ipynb_checkpoints"]
 
 today_fmt = "%Y-%m-%d %H:%M"
 
-# html_theme = "alabaster"
-# html_theme = "furo"
-html_theme = "pydata_sphinx_theme"
+html_css_files = [
+    "css/custom.css",
+]
 html_static_path = ["_static"]
+html_theme = "pydata_sphinx_theme"
+# fmt: off
+html_theme_options = {
+    "navbar_start": ["navbar-logo", "version-switcher"],
+    "switcher": {
+        "json_url": switcher_json_url,
+        "version_match": release if release in switcher_version_list else "dev"
+    }
+}
+# fmt: on
 
 autodoc_mock_imports = """
     bluesky